blockchain/requirements.txt
aptos-sdk>=1.7.0
web3>=6.11.1
python-dotenv>=1.0.0
eth-account>=0.8.0
eth-typing>=3.5.0
eth-utils>=2.3.0
requests>=2.31.0
pycryptodome>=3.19.0 